% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.r
\name{atlas}
\alias{atlas}
\title{Atlas creation}
\usage{
atlas(bedmatrix, segments.list, n, min.quality, epsilon = 0.001)
}
\arguments{
\item{bedmatrix}{a bed.matrix}

\item{segments.list}{a list of segments used to create the submaps}

\item{n}{the number of submaps}

\item{min.quality}{minimal quality (in \%) to include an inbred individual into the analysis}

\item{epsilon}{genotype error rate (default is 0.001)}
}
\description{
This function construct and complete the atlas
}
