% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.HBD.plot.r
\name{glm.HBD.plot}
\alias{glm.HBD.plot}
\title{Manhattan Plot for glm on HBD probabilities or FLOD}
\usage{
glm.HBD.plot(
  x,
  test = c("bilateral", "right", "left"),
  chrom.col = c("darksalmon", "darkturquoise"),
  ...
)
}
\arguments{
\item{x}{a data frame such as sent by HBD.glm}

\item{test}{which test to plot}

\item{chrom.col}{chromosome colors}

\item{...}{extra arguments for the plot}
}
\description{
Manhattan Plot for glm on HBD probabilities or FLOD
}
\seealso{
\code{\link{HBD.glm}}
}
