% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBD.plot.id.r
\name{HBD.plot.id}
\alias{HBD.plot.id}
\title{Plot the HBD segments of a given individual}
\usage{
HBD.plot.id(atlas, ROH, unit = "cM", id, famid, regions, outfile, build = 37)
}
\arguments{
\item{atlas}{an atlas object}

\item{ROH}{(optional) a data frame from which the segments will be plotted}

\item{unit}{the unit used to plot, "Bases" or "cM" (default is "cM")}

\item{id}{the individual id of the individual wanted}

\item{famid}{the family id of the individual wanted}

\item{regions}{a specific region to be highlighted in the plot (optional)}

\item{outfile}{a name for the plot (optional)}

\item{build}{the value of the genome build to use to plot chromosome in the plot (35, 36, 37,or 38, default is 37)}
}
\value{
A plot of the individual's HBDsegments.
}
\description{
This function plots all the HBD segments for all chromosomes and a given individual
}
\details{
If `ROH` is provided, it must be a data frame containing columns `FID`, `IID`, `CHR`, `POS1`, `POS2`, `PHE`. 
Such a data frame can be obtained by reading the report generated by `plink` with the option `--homozyg`.
If `ROH` is not provided, the function will plot HBD segments defined in the atlas object.

The `regions` argument is optional. If provided, it must be a matrix containing one line per region to be highlighted with in each line : 
\itemize{ 
  \item{the chromosome number}
  \item{start}
  \item{end}
}
}
\examples{
 
#Please refer to vignette 

}
