% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBD.segments.r
\name{HBD.segments}
\alias{HBD.segments}
\title{Find HBD segments}
\usage{
HBD.segments(x, n.consecutive.markers = 5, threshold = 0.5)
}
\arguments{
\item{x}{a atlas object}

\item{n.consecutive.markers}{the number of consecutive markers with a probabilitie equal or greater to the value of the threshold, to be use to fing HBD segments (default is 5)}

\item{threshold}{the minimum value of HBD probabilities given for a marker (default is 0.5)}
}
\value{
This function returns a list of dataframe with 11 columns :

- id

- famid

- pheno

- start

- end

- size

- chromosome

- start_pos

- end_pos

- start_dist

- end_dist
}
\description{
This function creates HBD segments used for creating plots
}
\details{
The threshold is the minimum value from which we consider the marker is HBD. From this marker we want a minumum number of consecutive markers to create a HBD segment.
(argument : n.consecutive.markers)
}
\keyword{internal}
