% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFLOD.manhattan.plot.r
\name{HFLOD.manhattan.plot}
\alias{HFLOD.manhattan.plot}
\title{Creation of an manhattan plot of the HFLOD}
\usage{
HFLOD.manhattan.plot(HFLOD, regions, unit = "cM", MA = FALSE, nbSNP_MA = 50)
}
\arguments{
\item{HFLOD}{the output of the HBD.gwas function}

\item{regions}{a matrix containing the value to be highlighted in the plot}

\item{unit}{the unit used to plot, two options are allowed "Bases", "cM" (default is "cM")}

\item{MA}{a boolean indicating whether a red line has to be drawn for the moving average}

\item{nbSNP_MA}{number of SNP for the moving average (default is 50)}
}
\value{
This function returns a manhattan plot of all the HFLOD score over all the chromosome
}
\description{
This fonction plot a manhanttan plot of the HFLOD score
}
\details{
If you use the regions options make sure to pass a matrix containing one line per region to be highlighted with in each line :

- the chromosome number

- start

- end
}
\examples{
#Please refer to vignette 

}
\seealso{
setHFLOD
}
