% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-class.r
\name{initialize,atlas-method}
\alias{initialize,atlas-method}
\title{Constructor of class atlas.}
\usage{
\S4method{initialize}{atlas}(
  .Object,
  bedmatrix,
  seeds,
  epsilon,
  segments.list,
  estimations,
  submap.summary
)
}
\arguments{
\item{.Object}{the object}

\item{bedmatrix}{a bed.matrix object}

\item{segments.list}{a list of segments object}
}
\description{
Constructor of class atlas.
}
