% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.r
\name{likelihoods}
\alias{likelihoods}
\title{Compute the likelihood for a given f and a parameters}
\usage{
likelihoods(atlas, f, a)
}
\arguments{
\item{atlas}{an atlas object}

\item{f}{a value of inbreeding coefficient f}

\item{a}{a value of HBD length parameter a}
}
\description{
Compute the likelihood for a given f and a parameters
}
