% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.r
\name{restore,atlas-method}
\alias{restore,atlas-method}
\title{Restore HBD and FLOD matrices}
\usage{
\S4method{restore}{atlas}(object)
}
\arguments{
\item{atlas}{an atlas}
}
\value{
an atlas
}
\description{
if atlas was built with memory mapped objects for HBD and FLOD matrices, 
and their pointers is broken, this method will attempt to recreate a valid pointer, if the 
file still exists.
}
