% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.list.by.hotspots.r
\name{segments.list.by.hotspots}
\alias{segments.list.by.hotspots}
\title{Creation of a list of segments}
\usage{
segments.list.by.hotspots(
  bedmatrix,
  intensity = 10,
  hotspots = hotspot_hg19,
  minMarkers = 0
)
}
\arguments{
\item{bedmatrix}{a bed.matrix object}

\item{intensity}{hotspots intensity threshold in cM/Mb (default is 10)}

\item{hotspots}{a data frame of recombination rates (default is hg19)}

\item{minMarkers}{minimum number of markers in a segment (default is 0)}
}
\value{
an hotspots.segments object
}
\description{
Creates a list of segments delimited by recombination hotspots
}
\details{
This function creates an object of class hotspots.segments, containing a list of segments delimited
by hotspots. The object is a list of list of vectors indices of SNPs. There are as many sublists as
chromosomes. The indices correspond to SNPs in \code{bedmatrix}.

The user can provide a hotspots data frame with a format similar to hotspot_hg19.
}
