% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.HFLOD.r
\name{set.HFLOD}
\alias{set.HFLOD}
\title{Computation of HFLOD scores}
\usage{
set.HFLOD(atlas, w.id)
}
\arguments{
\item{atlas}{a atlas object}

\item{w.id}{vector of indices to be considered in the FLOD_recap matrix}
}
\value{
the atlas object with its slot HFLOD completed
}
\description{
This function is used to compute HFLOD scores on individuals in a sample
}
