% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.pheno.r
\name{set.pheno}
\alias{set.pheno}
\title{Set phenotype}
\usage{
set.pheno(x, pheno)
}
\arguments{
\item{x}{a bedmatrix or an atlas}

\item{pheno}{a vector of phenotypes}
}
\value{
A bedmatrix or an atlas
}
\description{
Set phenotype
}
\details{
Update the phenotype in \code{x} and returns \code{x}
}
