% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-class.r
\name{show,atlas-method}
\alias{show,atlas-method}
\title{Show method of class atlas.}
\usage{
\S4method{show}{atlas}(object)
}
\arguments{
\item{object}{an atlas object}
}
\description{
Show method of class atlas.
}
