% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submaps.summary.r
\name{submaps.summary}
\alias{submaps.summary}
\title{Submap summary}
\usage{
submaps.summary(bedmatrix, a, f, p.lrt, min.quality, a.threshold = 1)
}
\arguments{
\item{bedmatrix}{the original bedmatrix}

\item{a}{a matrix of the estimations of a for all individuals on each submaps}

\item{f}{a matrix of the estimations of f for all individuals on each submaps}

\item{p.lrt}{a matrix of the pvalue of the likelihood test for all individuals on each submaps}

\item{min.quality}{minimal quality (in \%) to include an inbred individual into the analysis (default is 95)}

\item{a.threshold}{the maximum value for a (default is 1)}
}
\value{
this function returns a dataframe.
}
\description{
This function creates a summary on the submaps created.
}
\details{
This function gives for each genotyped individual summary statistics about the calculations.

This function returns a dataframe with 10 columns :

- famid : family identifier

- id : individual identifier

- pheno : phenotype (1 non affected, 2 affected, 0 unknown)

- submaps : number of valid submaps (i.e. submaps with a < a.threshold)

- quality: percentage of valid submaps

- f_median: median f on valid submaps (recommended to estimate f)

- a_median: median a on valid submaps (recommended to estimate a)

- pLRT_median: median p-value of LRT tests on valid submaps

- inbred: a flag indicating if the individual is inbred (pLRT_median <0.05) or not

- pLRT_<0.05: number of valid submaps with a LRT having a p-value below 0.05
}
\seealso{
setSummary
}
