% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matingtype.r
\name{T}
\alias{T}
\title{Mating type of individuals}
\usage{
T(alpha, L)
}
\description{
This function determines the mating type of the parents of all consanguineous individuals (1C,2C, 2x1C, AV or unrelated)
}
