% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.permutations.HFLOD.r
\name{threshold.permutations.HFLOD}
\alias{threshold.permutations.HFLOD}
\title{This function calculate the threshold of the association analysis with permutation method for the HBD-GWAS}
\usage{
threshold.permutations.HFLOD(
  atlas,
  nb.perm = 1000,
  phen,
  phen.code = c("R", "plink"),
  cores
)
}
\arguments{
\item{atlas}{an atlas object}
}
\description{
This function calculate the threshold of the association analysis with permutation method for the HBD-GWAS
}
