% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.r
\name{threshold}
\alias{threshold}
\title{Threshold of the association analysis}
\usage{
threshold(as, n.sims = 10000)
}
\arguments{
\item{as}{a data.frame with "chr", "dist" and "z.value" columns}

\item{n.sims}{the number of simulations (default is 1e4)}
}
\description{
This function calculate the threshold of the association analysis using a Ornstein-Uhlenbeck process
}
