% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fantasio.r
\name{Fantasio}
\alias{Fantasio}
\title{Wrapper for the package Fantasio}
\usage{
Fantasio(
  bedmatrix,
  segment.options,
  n = 100,
  min.quality = 95,
  allele.freq,
  q = 1e-04,
  epsilon = 0.001,
  epsilon2 = 0.001,
  median = TRUE,
  dense.recap = TRUE,
  basename
)
}
\arguments{
\item{bedmatrix}{a bed.matrix}

\item{segment.options}{a list of arguments to the function that will create the segments list}

\item{n}{the number of submaps (default is 100)}

\item{min.quality}{minimal quality (in \%) to include an inbred individual into the analysis (default is 95)}

\item{allele.freq}{a vector of allele frequencies (for allele A2), if \code{bedmatrix@p} isn't appropriate}

\item{q}{assumed frequency of the mutation involved in the disease for each individual (default is 0.0001)}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{epsilon2}{shift in border allele frequency p (p = 1 will be changed to 1-epsilon2 ; p = 0 will be changed to epsilon2)}

\item{median}{define the f and a parameters used to compute pHBD and FLOD
- if FALSE : f and a estimated on each submap
- if TRUE : median value of estimations on all submaps of f and a (default)}

\item{dense.recap}{define how pHBD and FLOD are combined over the submaps
- if FALSE : pHBD and FLOD are combined SNP by SNP with a mean on the number of submaps that includes this SNP
- if TRUE : mean of pHBD and FLOD on all submaps (default)}

\item{basename}{a basename for creating memory mapped objects (houba matrices)}
}
\description{
This function is used as a wrapper for the package Fantasio to create segments list and submaps
}
\details{
This function is a wrapper to make the usage of the package easier. The function calls different functions:

The first function, `segments.list.by.hotspots` is used to create a list of segments.

The second function, `atlas` is used to create submaps based on recombination hotspots.

The arguments that can be included in `segment.options` are described in `segments.list.by.hotspots`.
}
