% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fantasio.r
\name{Fantasio}
\alias{Fantasio}
\title{Wrapper for the package Fantasio}
\usage{
Fantasio(
  bedmatrix,
  segment.options,
  n = 100,
  min.quality = 95,
  list.id,
  recap = c("SNP", "segment"),
  phen.code = c("plink", "R"),
  q = 1e-04,
  epsilon = 0.001,
  median = TRUE,
  dense.recap = TRUE
)
}
\arguments{
\item{bedmatrix}{a bed.matrix}

\item{segment.options}{a list of arguments to the function that will create the segments list}

\item{n}{the number of submaps (default is 100)}

\item{min.quality}{minimal quality (in \%) to include an inbred individual into the analysis (default is 95)}

\item{list.id}{a list of individuals of interest ('famid:id') (default = no list id)}

\item{recap}{if you want the summary of probabilities by snps or by segments (only by SNPs for the moment)}

\item{phen.code}{phenotype coding :
- 'R' : 0:control ; 1:case ; NA:unknown 
- 'plink' : 1:control ; 2:case ; 0/-9/NA:unknown (default)}

\item{q}{assumed frequency of the mutation involved in the disease for each individual (default is 0.0001)}

\item{epsilon}{genotype error rate (default is 0.001)}

\item{median}{define the f and a parameters used to compute pHBD and FLOD
- if FALSE : f and a estimated on each submap
- if TRUE : median value of estimations on all submaps of f and a (default)}

\item{dense.recap}{define how pHBD and FLOD are combined over the submaps
- if FALSE : pHBD and FLOD are combined SNP by SNP with a mean on the number of submaps that includes this SNP
- if TRUE : mean of pHBD and FLOD on all submaps (default)}
}
\description{
This function is used as a wrapper for the package Fantasio to create segments list and submaps
}
\details{
This function is a wrapper to make the usage of the package easier. The function calls different functions:

The first function, `segments.list.by.hotspots` is used to create a list of segments.

The second function, `atlas` is used to create submaps based on recombination hotspots (for the moment).

The arguments that can be included in `segment.options` are described in `segments.list.by.hotspots`.

If `recap = 'SNP'`, the quantities such as HBD probabilities, FLOD, HFLOD, 
  are recapitulated SNP by SNP (default).
}
