% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.HBD.r
\name{glm.HBD}
\alias{glm.HBD}
\title{Logistic regression on HBD probability or FLOD score}
\usage{
glm.HBD(
  x,
  expl_var,
  covar_df,
  covar,
  test = match.arg("bilateral", "right", "left"),
  run = FALSE,
  phen.code
)
}
\arguments{
\item{x}{an atlas object}

\item{expl_var}{the explanatory variable 'FLOD' or 'pHBD'}

\item{covar_df}{a dataframe or a matrix containing covariates}

\item{covar}{covariates of interest such as 'age', 'sex' , ...
if missing, all covariates of the dataframe are considered}

\item{test}{which test to run. Default is bilateral. Use \code{"right"} to test for deleterious effects.}

\item{run}{whether the fonction is called or not (default = FALSE)}

\item{phen.code}{phenotype coding :
       - 'R' : 0:control ; 1:case ; NA:unknown (default)
       - 'plink' : 1:control ; 2:case ; 0/-9/NA:unknown
if 'plink' the function automatically convert it to 'R' to run logistic regression}

\item{n.cores}{number of cores for parallelization calculation (default = 1)}
}
\description{
Logistic regression on HBD probability or FLOD score
}
