% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.HBD.FLOD.sparse.r
\name{recap.HBD.FLOD.sparse}
\alias{recap.HBD.FLOD.sparse}
\title{Sparse summary of pHBD and FLOD}
\usage{
recap.HBD.FLOD.sparse(atlas, keep.inds, q, recap, median)
}
\arguments{
\item{atlas}{an atlas object}

\item{keep.inds}{the vector of index of consanguineous individuals}

\item{q}{assumed frequency of the mutation involved in the disease for each individual (default is 0.0001)}

\item{recap}{if you want the summary of probabilities by snps or by segments (only by SNPs for the moment)}

\item{median}{define the f and a parameters used to compute pHBD and FLOD
- if FALSE : f and a estimated on each submap
- if TRUE : median value of estimations on all submaps of f and a (default)}
}
\description{
pHBD and FLOD are combined over the submaps SNP by SNP with a mean on the number of submaps that includes this SNP
}
