% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.permutations.r
\name{threshold.permutations}
\alias{threshold.permutations}
\title{This function calculate the threshold of the association analysis with permutation method}
\usage{
threshold.permutations(
  atlas,
  nb.perm = 1000,
  expl.var = c("FLOD", "pHBD"),
  phen,
  phen.code = c("R", "plink"),
  covar_df = NULL,
  covar = NULL,
  score,
  centered = TRUE,
  cores
)
}
\arguments{
\item{atlas}{an atlas object}
}
\description{
This function calculate the threshold of the association analysis with permutation method
}
